+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Fixed
 - Config Auto-reload is now a repeating task as intended
+- Config actually created on first run
+- Copy updated default header when config is updated
 
 ## [1.0.0] - 2020-02-10
 ### Added